Server/requirements.txt
astroid==3.3.9
blinker==1.9.0
click==8.1.8
coverage==7.6.12
DateTime==5.5
dill==0.3.9
dotenv==0.9.9
Flask==3.1.0
greenlet==3.1.1
gunicorn==23.0.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
packaging==24.2
platformdirs==4.3.6
playwright==1.50.0
psycopg2-binary==2.9.10
pyee==12.1.1
pylint==3.3.5
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2025.2
sendgrid==6.11.0
setuptools==80.0.0
starkbank-ecdsa==2.2.0
tomlkit==0.13.2
typing_extensions==4.12.2
Werkzeug==3.1.3
zope.interface==7.2
